Coded the Show Details fragment

+++ app/src/main/java/ait/cativoapp/FavoriteClickListener.java
@@ -4,6 +4,7 @@ package ait.cativoapp;
 import android.content.Context;
 import android.view.View;
 import android.widget.ImageButton;
+import android.widget.LinearLayout;
 import android.widget.Toast;
 
 /**
@@ -13,8 +14,10 @@ import android.widget.Toast;
 public class FavoriteClickListener implements View.OnClickListener
 {
     SearchResult result;
+    ShowDetails details;
     ImageButton starButton;
     Context context;
+    LinearLayout progressBarLayout;
 
     public FavoriteClickListener(SearchResult result, ImageButton starButton, Context context )
     {
@@ -23,34 +26,79 @@ public class FavoriteClickListener implements View.OnClickListener
         this.context = context;
     }
 
+    public FavoriteClickListener(ShowDetails details, ImageButton starButton, Context context, View view )
+    {
+        this.details = details;
+        this.starButton = starButton;
+        this.context = context;
+        progressBarLayout = view.findViewById(R.id.show_details_watchedPBLayout);
+    }
+
     @Override
     public void onClick(View view)
     {
-        //if(starButton.getDrawable().getConstantState() == context.getResources().getDrawable(R.drawable.star_filled_white,context.getTheme()).getConstantState())
-        if(!result.isFavorite())
+        if(result != null)
         {
-            try
+            if (!result.isFavorite())
             {
-                DB.AddSerie(result.getId());
-                starButton.setImageResource(R.drawable.star_filled_yellow);
-                result.setFavorite(true);
+                try
+                {
+                    DB.AddSerie(result.getId());
+                    starButton.setImageResource(R.drawable.star_filled_yellow);
+                    result.setFavorite(true);
+                }
+                catch (Exception e)
+                {
+                    Toast.makeText(context, "Star Button selection error", Toast.LENGTH_LONG)
+                         .show();
+                }
             }
-            catch (Exception e)
+            else
             {
-                Toast.makeText(context, "Star Button selection error", Toast.LENGTH_LONG).show();
+                try
+                {
+                    DB.DeleteSerie(result.getId());
+                    starButton.setImageResource(R.drawable.star_filled_white);
+                    result.setFavorite(false);
+                }
+                catch (Exception e)
+                {
+                    Toast.makeText(context, "Star Button deselection error", Toast.LENGTH_LONG)
+                         .show();
+                }
             }
         }
-        else
+        if(details != null)
         {
-            try
+            if (!details.isFavorite())
             {
-                DB.DeleteSerie(result.getId());
-                starButton.setImageResource(R.drawable.star_filled_white);
-                result.setFavorite(false);
+                try
+                {
+                    DB.AddSerie(details.getId());
+                    starButton.setImageResource(R.drawable.star_filled_yellow);
+                    details.setFavorite(true);
+                    progressBarLayout.setVisibility(View.VISIBLE);
+                }
+                catch (Exception e)
+                {
+                    Toast.makeText(context, "Star Button selection error", Toast.LENGTH_LONG)
+                         .show();
+                }
             }
-            catch (Exception e)
+            else
             {
-                Toast.makeText(context, "Star Button deselection error", Toast.LENGTH_LONG).show();
+                try
+                {
+                    DB.DeleteSerie(details.getId());
+                    starButton.setImageResource(R.drawable.star_filled_white);
+                    details.setFavorite(false);
+                    progressBarLayout.setVisibility(View.GONE);
+                }
+                catch (Exception e)
+                {
+                    Toast.makeText(context, "Star Button deselection error", Toast.LENGTH_LONG)
+                         .show();
+                }
             }
         }
